pyproject.toml
[project]
name = "traffic-violation-system"
version = "0.1.0"
description = "Edge-First Smart Traffic Violation Detection System using YOLO26n + OpenVINO"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Muhammad Thabsheer" },
]
keywords = ["edge-ai", "traffic", "computer-vision", "yolo", "openvino", "smart-city"]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy>=2.0.0",
    "pydantic-settings>=2.0.0",
    "openvino>=2024.6",
    "ultralytics>=8.3.0",
    "opencv-python-headless>=4.10.0",
    "numpy>=1.26.0",
    "scipy>=1.14.0",
    "nncf>=2.14.0",
    "websockets>=13.0",
    "python-multipart>=0.0.12",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["backend*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py311"
line-length = 100
line-ending = "lf"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501", "E402"]
